refactor: 코멘트 반영
