+++ src/main/java/ddingdong/ddingdongBE/common/exception/CustomException.java
@@ -1,8 +1,11 @@
 package ddingdong.ddingdongBE.common.exception;
 
+import lombok.Getter;
+
+@Getter
 abstract class CustomException extends RuntimeException {
-    String message;
-    int errorCode;
+    private final String message;
+    private final int errorCode;
 
     public CustomException(String message, int errorCode) {
         super(message);
