+++ src/main/java/ddingdong/ddingdongBE/domain/club/controller/dto/request/RegisterClubRequest.java
@@ -29,8 +29,8 @@ public class RegisterClubRequest {
                 .category(category)
                 .tag(tag)
                 .leader(leaderName)
-                .location(Location.of("S0000"))
-                .phoneNumber(PhoneNumber.of("010-0000-0000"))
+                .location(Location.from("S0000"))
+                .phoneNumber(PhoneNumber.from("010-0000-0000"))
                 .score(Score.from(0)).build();
     }
